// src/types/smtrack/repair/repairType.ts
type RepairType = {
  address: string
  createAt: string
  detail: string
  devName: string
  device: {
    id: string
    name: string
  }
  id: string
  info: string
  info1: string
  info2: string
  phone: string
  remark: string
  seq: number
  status: string
  updateAt: string
  ward: string
  warrantyStatus: string
}

export type { RepairType }